@@ -1,7 +1,7 @@
 package com.example.restexample
 
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.widget.ImageView
 import android.widget.Toast
 import com.squareup.picasso.Picasso
